[level 0] Title: 두 수의 나눗셈, Time: 0.02 ms, Memory: 88.1 MB -BaekjoonHub

+++ 프로그래머스/0/120806. 두 수의 나눗셈/두 수의 나눗셈.java
@@ -1,10 +1,10 @@
 class Solution {
-    public int solution(float num1, float num2) {
+    public int solution(int num1, int num2) {
         int answer = 0;
         if ((0 < num1 && 100 >= num1) &&
             (0 < num2 && 100 >= num2)) {
             
-           float result = (num1 / num2) * 1000.0f;
+            double result = ((double) num1 / (double) num2) * 1000.0;
             answer = (int) result;
         }
         return answer;
